+++ src/App.js
@@ -1,6 +1,6 @@
 import './App.css';
 import Navbar from './Components/Navbar';
-import Home from './Pages/Home';
+import Home from './Pages/Home/Home';
 
 function App() {
   return (
